+++ partials/specials/banners/podcast-committed.php
@@ -1,6 +1,6 @@
 <?php
   $local_term = get_term_by( 'slug', 'committed', 'category' );
-  $base_image_path = get_stylesheet_directory_uri() . '/dist/img/specials/committed/';
+  $base_image_path = get_stylesheet_directory_uri() . '/src/img/specials/committed/';
   $url = $local_term ? get_term_link( $local_term ) : false;
 if ( $local_term && $url ) {
   ?>
@@ -78,13 +78,19 @@ if ( $local_term && $url ) {
   <div class="container committed-banner__container pt-6 pb-8 p-s-4">
     <div class="grid-row committed-banner__grid-row">
       <div class="grid-item is-s-12 is-xxl-6 is-m-7 committed-banner__grid-item committed-banner__grid-item-left mt-s-4">
-        <p class="font-color-white font-weight-bold font-size-11 font-size-s-11">Would you go to prison for your politics?</p>
+        <a href="<?php echo esc_url( $url ); ?>" class="text-link-wrapper">
+          <p class="font-color-white font-weight-bold font-size-11 font-size-s-11">Would you go to prison for your politics?</p>
+        </a>
       </div>
       <div class="grid-item is-s-24 is-xs-20 is-xxl-12 is-m-10 committed-banner__grid-item committed-banner__logo">
       </div>
       <div class="grid-item is-s-12 is-xxl-6 is-m-7 committed-banner__grid-item committed-banner__grid-item-right mt-s-4">
         <div class="committed-banner__button-container committed-banner__grid-item-right">
-          <p class="font-color-white font-weight-bold font-size-11 mb-3 pt-6 pt-s-0 font-size-s-11">Meet the people who</br>have done just that.</p>
+          <p class="font-color-white font-weight-bold font-size-11 mb-3 pt-6 pt-s-0 font-size-s-11">
+            <a href="<?php echo esc_url( $url ); ?>" class="text-link-wrapper">
+              Meet the people who<br>have done just that.
+            </a>
+          </p>
           <a href="<?php echo esc_url( $url ); ?>" class="ui-button ui-button--small ui-button--white">Listen Now</a>
         </div>
       </div>